feat: rename helpers folder
