+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './helper/index.js';
+export * from './helpers/index.js';
 export * from './middleware/index.js';
 export * from './reactive/index.js';
 export * from './types/index.js';
