+++ src/reactive/pipe.ts
@@ -1,7 +1,7 @@
 import { fold } from 'fp-ts/Either';
 import type { Union } from 'ts-toolbelt';
-import { isPromise } from '../helper/helper-is-promise.js';
-import { pipeWith } from '../helper/helper-pipe-with.js';
+import { isPromise } from '../helpers/helper-is-promise.js';
+import { pipeWith } from '../helpers/helper-pipe-with.js';
 import {
   Exists,
   ExitPipeReturnValue,
